basic attacks now kind of do their thing!
When user clicks on an enemy, they now attack it only if the player
has at least 1 Basic Attack.
The player still can play the basic attack as any card, but this
will not do anything for now. ideally this should highlight all
entities which can be targeted by it and let the user click/choose
one of these. if only 1 targetable entity exists, it should be
targeted automatically.

 * added: card types (basic-attack or action, for now...)
 * added: player needs to have a basic attack to attack
 * added: also allow adding handcards during player turn, due to card effect to draw a card
 * fixed: flecs ecs_filter_fini sometimes not called?